pyproject.toml
[tool.poetry]
name = "JackpotMaximize"
version = "0.1.0"
description = "JackpotMaximize S2Q"
authors = ["JogfeeingVI <lifelse@outlook.com>"]
license = "GLP3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
toml = "^0.10.2"
ipykernel = "^6.21.1"
setuptools = "^68.2.2"
gprof2dot = "^2022.7.29"
pytest = "^7.4.4"
aiosqlite = "^0.20.0"
emoji = "^2.11.0"
httpx = "^0.27.0"
pillow = "^10.3.0"
socksio = "^1.0.0"
tqdm = "^4.66.5"


[tool.poetry.group.dev.dependencies]


[tool.poetry.group.new_group]
optional = true

[tool.poetry.group.new_group.dependencies]
tensorflow = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
